FAT2-333 - Change to provider ordering
Make sure that At Work Place only are grouped together with similar scores and not put into one group on their own.
Split out extension to do scoring, then grouping on distance by score. Created separate test file.
