+++ src/SFA.DAS.FindApprenticeshipTraining.Api.UnitTests/Extensions/WhenOrderingByProviderRating.cs
@@ -1,9 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using AutoFixture.NUnit3;
 using FluentAssertions;
 using NUnit.Framework;
-using NUnit.Framework.Internal;
 using SFA.DAS.FindApprenticeshipTraining.Api.Extensions;
 using SFA.DAS.FindApprenticeshipTraining.Api.Models;
 
@@ -59,7 +57,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
             
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(4);
             list.Skip(1).Take(1).First().ProviderId.Should().Be(3);
@@ -100,7 +98,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
 
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(3);
         }
@@ -166,7 +164,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
 
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(3);
             list.Last().ProviderId.Should().Be(1);
@@ -243,7 +241,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
 
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(3);
             list.Last().ProviderId.Should().Be(1);
@@ -361,7 +359,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
 
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(3);
             list.Last().ProviderId.Should().Be(1);
@@ -401,7 +399,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
             
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(4);
             list.Skip(1).Take(1).First().ProviderId.Should().Be(1);
@@ -448,7 +446,7 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
             
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(4);
             list.Skip(1).Take(1).First().ProviderId.Should().Be(1);
@@ -495,635 +493,13 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.UnitTests.Extensions
                 }
             };
             
-            list = list.OrderByProviderRating().ToList();
+            list = list.OrderByProviderScore().ToList();
 
             list.First().ProviderId.Should().Be(3);
             list.Skip(1).Take(1).First().ProviderId.Should().Be(4);
             list.Last().ProviderId.Should().Be(1);
         }
         
-        [Test]
-        public void Then_If_There_Is_A_Location_Then_Scored_And_Sorted_By_Distance()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 0
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 2,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 1
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 5.1m
-                        }
-                    } ,
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 4,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 15.1m
-                        }
-                    },
-                    HasLocation = true
-                }
-            };
-            
-            list = list.OrderByProviderRating().ToList();
-            
-            list.First().ProviderId.Should().Be(1);
-            list.Skip(1).Take(1).First().ProviderId.Should().Be(3);
-            list.Skip(2).Take(1).First().ProviderId.Should().Be(2);
-            list.Last().ProviderId.Should().Be(4);
-        }
         
-        [Test]
-        public void Then_Delivery_Type_Of_Workplace_Counts_As_Zero_Distance()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 0
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 2,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 1
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 5.1m
-                        }
-                    } ,
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 4,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 15.1m
-                        }
-                    },
-                    HasLocation = true
-                }
-            };
-            
-            list = list.OrderByProviderRating().ToList();
-            
-            list.First().ProviderId.Should().Be(3);
-            list.Skip(1).Take(1).First().ProviderId.Should().Be(4);
-            list.Skip(2).Take(1).First().ProviderId.Should().Be(1);
-            list.Last().ProviderId.Should().Be(2);
-        }
-        
-        [Test]
-        public void Then_If_Matched_In_Group_Then_Ordered_By_Closest_Excluding_Workplace_If_Other_Delivery_Options()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 3.9m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 0m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 2,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 0m
-                        }
-                    },
-                    HasLocation = true
-                }
-            };
-            
-            list = list.OrderByProviderRating().ToList();
-            
-            list.First().ProviderId.Should().Be(3);
-            list.Skip(1).First().ProviderId.Should().Be(1);
-            list.Last().ProviderId.Should().Be(2);
-        }
-
-        [Test]
-        public void Then_If_There_Are_Multiple_Filters_Then_Closest_Is_Used()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 1.2m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 1.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 3m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 2,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.8m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 0m
-                        }
-                    },
-                    HasLocation = true
-                }
-            };
-            
-            list = list.OrderByProviderRating(new List<DeliveryModeType> {DeliveryModeType.DayRelease, DeliveryModeType.BlockRelease}).ToList();
-            
-            list.First().ProviderId.Should().Be(2);
-            list.Skip(1).First().ProviderId.Should().Be(1);
-            list.Last().ProviderId.Should().Be(3);
-        }
-        
-        [Test]
-        public void Then_If_There_Is_A_Delivery_Filter_For_DayRelease_Then_That_Mode_Is_Used_For_Distance()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 1.2m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 1.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 3m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 2,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.8m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 0m
-                        }
-                    },
-                    HasLocation = true
-                }
-            };
-            
-            list = list.OrderByProviderRating(new List<DeliveryModeType> {DeliveryModeType.DayRelease}).ToList();
-            
-            list.First().ProviderId.Should().Be(3);
-            list.Skip(1).First().ProviderId.Should().Be(2);
-            list.Last().ProviderId.Should().Be(1);
-        }
-        
-        [Test]
-        public void Then_If_There_Is_A_Delivery_Filter_For_Block_Release_Then_That_Mode_Is_Used_For_Distance()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 1.2m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 1.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 3m
-                        }
-                    },
-                    HasLocation = true
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 2,
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 4
-                    },
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.DayRelease,
-                            DistanceInMiles = 4.9m
-                        },
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 0m
-                        }
-                    },
-                    HasLocation = true
-                }
-            };
-            
-            list = list.OrderByProviderRating(new List<DeliveryModeType> {DeliveryModeType.BlockRelease}).ToList();
-            
-            list.First().ProviderId.Should().Be(2);
-            list.Skip(1).First().ProviderId.Should().Be(1);
-            list.Last().ProviderId.Should().Be(3);
-        }
-        
-        [Test]
-        public void Then_If_All_Are_At_Workplace_Then_Not_Grouped_By_Distance()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Name = "Third",
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3,
-                        TotalEmployerResponses = 100
-                    },
-                    OverallCohort = 100,
-                    HasLocation = true,
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 10
-                        }
-                    }
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 4,
-                    Name = "Second",
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3,
-                        TotalEmployerResponses = 100
-                    },
-                    OverallCohort = 100,
-                    HasLocation = true,
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 15
-                        }
-                    }
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Name = "First",
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3,
-                        TotalEmployerResponses = 100
-                    },
-                    OverallCohort = 100,
-                    HasLocation = true,
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.Workplace,
-                            DistanceInMiles = 100
-                        }
-                    }
-                }
-            };
-            
-            list = list.OrderByProviderRating().ToList();
-
-            list.First().ProviderId.Should().Be(3);
-            list.Skip(1).Take(1).First().ProviderId.Should().Be(4);
-            list.Last().ProviderId.Should().Be(1);
-        }
-        
-        [Test]
-        public void Then_If_All_Are_Same_Score_And_Distance_Then_Ordered_By_OverAllCohort()
-        {
-            var list = new List<GetTrainingCourseProviderListItem>
-            {
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 1,
-                    Name = "Test",
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3,
-                        TotalEmployerResponses = 100
-                    },
-                    OverallCohort = 10,
-                    HasLocation = true,
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 11
-                        }
-                    }
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 4,
-                    Name = "Test",
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3,
-                        TotalEmployerResponses = 100
-                    },
-                    OverallCohort = 100,
-                    HasLocation = true,
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 11
-                        }
-                    }
-                },
-                new GetTrainingCourseProviderListItem
-                {
-                    ProviderId = 3,
-                    Name = "Test",
-                    Feedback = new GetProviderFeedbackResponse
-                    {
-                        TotalFeedbackRating = 3,
-                        TotalEmployerResponses = 100
-                    },
-                    OverallCohort = 1000,
-                    HasLocation = true,
-                    DeliveryModes = new List<GetDeliveryType>
-                    {
-                        new GetDeliveryType
-                        {
-                            DeliveryModeType = DeliveryModeType.BlockRelease,
-                            DistanceInMiles = 11
-                        }
-                    }
-                }
-            };
-            
-            list = list.OrderByProviderRating().ToList();
-
-            list.First().ProviderId.Should().Be(3);
-            list.Skip(1).Take(1).First().ProviderId.Should().Be(4);
-            list.Last().ProviderId.Should().Be(1);
-        }
     }
 }
