+++ src/SFA.DAS.FindApprenticeshipTraining.Api/Models/GetTrainingCourseResponse.cs
@@ -4,6 +4,5 @@ namespace SFA.DAS.FindApprenticeshipTraining.Api.Models
     {
         public GetTrainingCourseListItem TrainingCourse { get; set; }
         public GetTrainingCourseProviderCountResponse ProvidersCount { get; set; }
-        public string LevelEquivalent { get; set; }
     }
 }
